Move helpers into own file
